requirements.txt
langchain
transformers
torch
chromadb
pypdf2
python-docx
fastapi
uvicorn
pydantic
PyYAML
accelerate
huggingface-hub
protobuf
einops
safetensors
tokenizers
python-multipart